+++ packages/online-editor/src/accounts/authSessions/AuthSessionLabel.tsx
@@ -16,7 +16,7 @@
 
 import * as React from "react";
 import { Text, TextContent, TextVariants } from "@patternfly/react-core/dist/js/components/Text";
-import { AuthSession } from "./AuthSessionsContext";
+import { AuthSession } from "./AuthSessionApi";
 import { useAuthProvider } from "../authProviders/AuthProvidersContext";
 import { AuthProviderIcon } from "../authProviders/AuthProviderIcon";
 import { Flex } from "@patternfly/react-core/dist/js/layouts/Flex";
