+++ packages/online-editor/src/importFromUrl/ImportableUrlHooks.tsx
@@ -27,9 +27,11 @@ import { useCallback, useMemo } from "react";
 import { matchPath } from "react-router";
 import { AuthProviderIcon } from "../accounts/authProviders/AuthProviderIcon";
 import { useAuthProvider, useAuthProviders } from "../accounts/authProviders/AuthProvidersContext";
-import { AuthInfo, AuthSession } from "../accounts/authSessions/AuthSessionsContext";
+import { AuthSession } from "../accounts/authSessions/AuthSessionApi";
+import { AuthInfo } from "../accounts/authSessions/AuthSessionsContext";
 import { useEditorEnvelopeLocator } from "../envelopeLocator/hooks/EditorEnvelopeLocatorContext";
-import { getGitRefName, getGitRefType, GitRefTypeIcon } from "../gitRefs/GitRefs";
+import { getGitRefName, getGitRefType } from "../gitRefs/GitRefs";
+import { GitRefTypeIcon } from "../gitRefs/GitRefTypeIcon";
 import { useSettings } from "../settings/SettingsContext";
 import { PromiseStateStatus, useLivePromiseState } from "../workspace/hooks/PromiseState";
 import { GitServerRef } from "../workspace/worker/api/GitServerRef";
